requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.13
# by the following command:
#
#    pip-compile requirements.in
#
aiohappyeyeballs==2.6.1
    # via aiohttp
aiohttp==3.12.13
    # via
    #   -r requirements.in
    #   python-socketio
aiosignal==1.3.2
    # via aiohttp
attrs==25.3.0
    # via aiohttp
bidict==0.23.1
    # via python-socketio
blinker==1.9.0
    # via flask
cffi==1.17.1
    # via gevent
click==8.2.1
    # via flask
colorama==0.4.6
    # via click
flask==3.1.1
    # via
    #   -r requirements.in
    #   flask-socketio
flask-socketio==5.5.1
    # via -r requirements.in
frozenlist==1.7.0
    # via
    #   aiohttp
    #   aiosignal
gevent==25.5.1
    # via
    #   -r requirements.in
    #   gevent-websocket
gevent-websocket==0.10.1
    # via -r requirements.in
greenlet==3.2.3
    # via gevent
h11==0.16.0
    # via wsproto
idna==3.10
    # via yarl
itsdangerous==2.2.0
    # via flask
jinja2==3.1.6
    # via flask
markupsafe==3.0.2
    # via
    #   flask
    #   jinja2
    #   werkzeug
multidict==6.6.3
    # via
    #   aiohttp
    #   yarl
propcache==0.3.2
    # via
    #   aiohttp
    #   yarl
pycparser==2.22
    # via cffi
python-engineio==4.12.2
    # via python-socketio
python-socketio[asyncio-client,asyncio_client]==5.13.0
    # via
    #   -r requirements.in
    #   flask-socketio
simple-websocket==1.1.0
    # via python-engineio
websocket-client==1.8.0
    # via -r requirements.in
werkzeug==3.1.3
    # via flask
wsproto==1.2.0
    # via simple-websocket
yarl==1.20.1
    # via aiohttp
zope-event==5.1
    # via gevent
zope-interface==7.2
    # via gevent

# The following packages are considered to be unsafe in a requirements file:
# setuptools
